requirements.txt
# Core
openai>=1.0.0
python-dotenv>=1.0.0

# LangChain ecosystem
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0

# Vector search
faiss-cpu>=1.7.4

# NLP / ML
torch>=2.0.0
transformers>=4.33.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# Web/Chatbot UI
gradio>=4.0.0

# Utilities
pydantic>=2.0.0
tqdm>=4.65.0
